[C++ SDK] Moved private executors to adapters
