+++ ydb/core/kqp/federated_query/ya.make
@@ -18,6 +18,7 @@ PEERDIR(
     ydb/library/yql/providers/s3/proto
     ydb/library/yql/providers/solomon/gateway
     ydb/public/api/protos
+    ydb/public/sdk/cpp/adapters/executor
     ydb/public/sdk/cpp/adapters/issue
     ydb/public/sdk/cpp/src/client/extensions/discovery_mutator
     yql/essentials/core/dq_integration/transform
