+++ ydb/public/sdk/cpp/adapters/executor/ya.make
@@ -1,0 +1,13 @@
+LIBRARY()
+
+SRCS(
+    executor.cpp
+)
+
+PEERDIR(
+    ydb/public/sdk/cpp/src/client/impl/executor
+    ydb/public/sdk/cpp/src/client/impl/internal/thread_pool
+    ydb/public/sdk/cpp/src/client/types/executor
+)
+
+END()
