+++ ydb/public/sdk/cpp/adapters/ya.make
@@ -1,3 +1,4 @@
 RECURSE(
+    executor
     issue
 )
