+++ ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/types/executor/executor.h
@@ -1,9 +1,5 @@
 #pragma once
 
-#ifndef YDB_SDK_OSS
-#include <util/thread/pool.h>
-#endif
-
 #include <functional>
 #include <memory>
 #include <mutex>
@@ -44,15 +40,5 @@ protected:
 
 // Create default executor for thread pool.
 IExecutor::TPtr CreateThreadPoolExecutor(std::size_t threadCount, std::size_t maxQueueSize = 0);
-
-#ifndef YDB_SDK_OSS
-// Create executor adapter for util thread pool.
-// Thread pool is started and stopped by SDK.
-IExecutor::TPtr CreateThreadPoolExecutorAdapter(std::shared_ptr<IThreadPool> threadPool, std::size_t threadCount, std::size_t maxQueueSize = 0);
-
-// Create executor adapter for util thread pool.
-// Thread pool is expected to have been started and stopped by user.
-IExecutor::TPtr CreateExternalThreadPoolExecutorAdapter(std::shared_ptr<IThreadPool> threadPool);
-#endif
 
 } // namespace NYdb
